% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroraffine.R
\name{apply.mirror.affine}
\alias{apply.mirror.affine}
\alias{apply.mirror.affine.default}
\alias{apply.mirror.affine.neuron}
\alias{apply.mirror.affine.neuronlist}
\title{Apply a mirroring and affine transformation to an object in 3D space}
\usage{
apply.mirror.affine(x, ...)

\method{apply.mirror.affine}{neuron}(x, ...)

\method{apply.mirror.affine}{neuronlist}(x, ...)

\method{apply.mirror.affine}{default}(x, calculatetransform = NULL,
  pathtomatrix = NULL, ...)
}
\arguments{
\item{x}{a set of 3D coordinates (Nx3 matrix), a neuron, neuronlist, dotprops, hxsurf, igraph or mesh3d object.
currently point replacement is only supported for neuron and neuronlist objects.}

\item{...}{additional arguments eventually passed to: \code{\link{calculate.full.transformation}} \code{\link{mirrormat}} \code{\link{transform3dpoints}}}

\item{calculatetransform}{whether or not to re-calculate the mirorring transformation, and the affine transformation for
the mirrored object onto the original ones. Null defaults to a list of transformation matrices generated from flipping a set
of neuroanatomical structures in the L1 larval central nervous system, then performing an affine transformation. Otherwise,}

\item{pathtomatrix}{path to a saved transformation matrix, or list of matrices, that you want to use}
}
\value{
a set of 3D cordinates (or a neuron/neuronlist object if that was given as input), that has undergone sequential mirroring
and affine transformations. The default is to flip points in the l1 larval central nervous system from one side to the
other.
}
\description{
Apply a mirroring and affine transformation to an object in 3D space
}
\seealso{
\code{\link{calculate.full.transformation}} \code{\link{mirrormat}} \code{\link{transform3dpoints}}
\code{\link{trafoicpmat}}
}

