% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroraffine.R
\name{calculate.full.transformation}
\alias{calculate.full.transformation}
\title{Calculate a mirroring, then a pair of affine transformations}
\usage{
calculate.full.transformation(objs = system.file("extdata/point_objects/",
  package = "deformetricar"), pattern = ".rds$", ...)
}
\arguments{
\item{objs}{An R object () or the location of (an) R object(s) for the basis of calculating the transformation matrices.
Defaults to a set of files describing key neuroanatomical structures of the L1 larval central nervous system.}

\item{pattern}{search term for RDS files to read in the given location.}

\item{...}{additional arguments eventually passed to: \code{\link{trafoicpmat}} \code{\link{mirrormat}}}
}
\value{
a list of transformation matrices
}
\description{
Calculates a mirroring transformation, then a rigid, then a non-rigid affine transformation on a given 3D R object, or
one(s) saved as RDS files By default, the mirroring occurs across the X axis of the provided data.
}

