% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformetricar.R
\docType{package}
\name{deformetricar-package}
\alias{deformetricar}
\alias{deformetricar-package}
\title{deformetricar: an R client for the Deformetrica shape-registration toolkit}
\description{
\bold{deformetricar} wraps the Deformetrica (>= 4.3) command-line tools for
statistical shape analysis: it fits diffeomorphisms between point clouds, neuron
backbones and surface meshes, and applies them by geodesic shooting. It also
reads and writes VTK object formats.
}
\section{Package options}{

\describe{
\item{\code{deformetricar.exe}}{Path to (or name of) the \code{deformetrica}
executable; consulted by \code{\link{find_deformetrica}}. Set it in your
\code{\link[base]{Rprofile}} if Deformetrica is not on the \code{PATH} or in
a conda environment.}
}
}

\references{
\url{https://www.deformetrica.org}
}
\seealso{
\code{\link{find_deformetrica}}, \code{\link{deformetrica_register}},
\code{\link{deformetrica_shoot}}
}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com}

Authors:
\itemize{
  \item Alexander Bates \email{alexander.shakeel.bates@gmail.com}
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
