% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpartner.R
\name{findcognate}
\alias{checkL1cognates}
\alias{findcognate}
\title{Find neuron cognates on other side of a nervous system.}
\usage{
findcognate(x, db, entries = 10, ...)

checkL1cognates(possiblecognates, db, entries = 10, ...)
}
\arguments{
\item{x}{a neuron or neuronlist from the Drosphila L1 connectome.}

\item{db}{a database of neurons against which to search for partners. Preferably a dotprops object.}

\item{entries}{number of possible ranked cognates to display}

\item{...}{additional arguments passed to: \code{\link{otherside}} Necessary if not using defauly nervous system fro this package, the larval L1 reconstruction by Cardona and colleagues.}

\item{possiblecognates}{object returned from calling findcognate}
}
\value{
"Scores" -  a matrix tabulating similarity scores (NBlast) for neurons on the otherside of the brain,
neurons being identified by skid, or "neuronlist" - a neuronlist containing the neurons with the highest similarity scores.
}
\description{
Find neuron cognates on other side of a nervous system.
}
\seealso{
\code{\link{shootflow}} \code{\link{apply.mirror.affine}} \code{\link{otherside}}
}

