% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpartner.R
\name{findpartner}
\alias{findpartner}
\alias{findpartner.default}
\title{Find neuron cognates on other side of a structure}
\usage{
findpartner(x, ...)

\method{findpartner}{default}(x, db, output = c("scores", "neuronlist"), ...)
}
\arguments{
\item{x}{a neuron or neuronlist from the Drosphila L1 connectome.}

\item{...}{additional arguments passed to: \code{\link{shootflow}} \code{\link{apply.mirror.affine}}}

\item{db}{a batabse of neurons against which to search for partners. Preferably a dotprops object.}
}
\value{
"Scores" -  a matrix tabulating similarity scores (NBlast) for neurons on the otherside of the brain,
neurons being identified by skid, or "neuronlist" - a neuronlist containing the neurons with the highest similarity scores.
}
\description{
Find neuron cognates on other side of a structure
}

