% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroraffine.R
\name{mirrormat}
\alias{mirrormat}
\title{Generate a transformation matrix for a mirroring registration}
\usage{
mirrormat(x, mirrorAxis = c("X", "Y", "Z"), mirrorAxisSize = NULL, ...)
}
\arguments{
\item{x}{a set of 3D coordinates (Nx3 matrix), a neuron, neuronlist, dotprops, hxsurf, igraph or mesh3d object.}

\item{mirrorAxis}{the axis along which to mirror the data. Defaults to the X axis.}

\item{mirrorAxisSize}{this defaults to the bounding box for x.}

\item{...}{additional arguments eventually passed to methods}
}
\value{
A transformation matrix
}
\description{
Generates
}

