% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherside.R
\name{otherside}
\alias{otherside}
\alias{otherside.default}
\alias{otherside.neuron}
\alias{otherside.neuronlist}
\title{Apply a list of mirroring, affine and warping registrations to an object in 3D space}
\usage{
otherside(x, method = c("saved", "tps3d", "deformetrica"), ...)

\method{otherside}{default}(x, method = c("saved", "tps3d", "deformetrica"),
  regdir = system.file("extdata/reg_output/", package = "deformetricar"),
  object.type = "NonOrientedPolyLine", ...)

\method{otherside}{neuron}(x, method = c("saved", "tps3d", "deformetrica"),
  object.type = "NonOrientedPolyLine", ...)

\method{otherside}{neuronlist}(x, method = c("saved", "tps3d",
  "deformetrica"), object.type = "NonOrientedPolyLine", ...)
}
\arguments{
\item{x}{a set of 3D coordinates (Nx3 matrix), a neuron, neuronlist, dotprops, hxsurf, igraph or mesh3d object.
currently point replacement is only supported for neuron and neuronlist objects.}

\item{method}{Whether to use the 3D thin plate spline mapping function from package Morpho, or, more accurately but much more slowly,
Deformetrica (must be installed separately) in order to recapitulate a Deformetricar deformation.}

\item{...}{additional arguments passed to: \code{\link{shootflow}} \code{\link{apply.mirror.affine}}}

\item{regdir}{Path to directory containing deformetrica registration.}

\item{object.type}{Type of object for Deformetrica deformation. See Deformetrica's documentation. Default is appropriate for neuron skeletons.}
}
\value{
a set of 3D cordinates (or a neuron/neuronlist object if that was given as input), that has undergone sequential mirroring
affine and warping (via Deformetrica, see http://www.deformetrica.org/?page_id=232) transformations. The default is to flip points in
the l1 larval central nervous system from one side to the other, and apply a previousy generated Deformetrica deformation based on the
registering of a set of flipped, large neuroanatomical regions unto their un-flipped counterparts.
}
\description{
Apply a list of mirroring, affine and warping registrations to an object in 3D space
}
\seealso{
\code{\link{shootflow}} \code{\link{apply.mirror.affine}}
}

