% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteVTK.R
\name{read.points}
\alias{read.points}
\title{Read a Deformetrica control-points / momenta txt file}
\usage{
read.points(filename)
}
\arguments{
\item{filename}{read CPS or MOM .txt file}
}
\value{
An N x 3 matrix.
}
\description{
Read a Deformetrica control-points / momenta txt file
}
\keyword{internal}
