% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteVTK.R
\name{read.vtk.to.mesh}
\alias{read.vtk.to.mesh}
\title{Read a mesh3d object from a .vtk file}
\usage{
read.vtk.to.mesh(filenames)
}
\arguments{
\item{filename}{a list of filenames to covnert into a single mesh3d object}
}
\description{
Read a mesh3d object from a .vtk file
}

