% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootflow.R
\name{shootflow}
\alias{shootflow}
\title{Apply deformetrica registration to a set of 3D points (ShootAndFlow3)}
\usage{
shootflow(points, regdir = getwd(), data.sigma = 1, kernel.width = 5,
  object.type = "NonOrientedPolyLine")
}
\arguments{
\item{points}{3D coordinates (Nx3 matrix)}

\item{regdir}{Path to directory containing deformetrica registration}

\item{data.sigma}{See deformetrica docs}

\item{kernel.width}{See deformetrica docs}

\item{object.type}{Type of object to be deformed}
}
\value{
Matrix of the same dimensions as \code{points}
}
\description{
Apply deformetrica registration to a set of 3D points (ShootAndFlow3)
}
\references{
See \url{http://www.deformetrica.org/?page_id=232} for
details of the \bold{ShootAndFlow3} command line tool.
}

