% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematching.R
\name{sparsematching3}
\alias{sparsematching3}
\title{Run a Deformatrica sparseMatching3 deformation on 3D objects}
\usage{
sparsematching3(regdir, choose.params = T, verbose = T, moving.list,
  target.list, k.width = 10, kernel.type = c("Exact", "cuda", "fgt", "p3m"),
  no.threads = 4, no.timepoints = 5, sparsity.prior = 0,
  max.iterations = 100, smooth.k.width.ratio = 1, data.sigma = 1,
  kernel.width = 10, data.type = "NoneOrientatedPolyline", plotting = F,
  app = sparseMatching3("sparseMatching3"), ...)
}
\arguments{
\item{regdir}{Path to directory for performing the deformetrica registration. The .vtk files produced for 3D objects and the registrations generated will be saved here}

\item{choose.params}{whether or not to manually select values for kernel.width, data.sigma and object.type for eaxch object in the registration. Else, default values will be used}

\item{verbose}{whether or not to see the progress of the sparseMatching3 system command in your R console}

\item{moving.list}{list of mesh3D objects to be deformed}

\item{target.list}{list of mesh3D objects the moving objects are goign to try to match}

\item{k.width}{the width of the deformation kernel for the whoel deformation. The larger the more
rigid the deformation. The smaller, the more local variations of the space
is allowed}

\item{kernel.type}{the kernel type used. See the deformetrica website for details}

\item{no.threads}{the number of thread to use on your machine during the calculation}

\item{no.timepoints}{the number of time stesp at which ti generate a .vtk files for the moving objects. The higher this value, the slower the calculation}

\item{smooth.k.width.ratio}{the degree of smoothening to kernel width ratio}

\item{data.sigma}{matching factor. The lower ir is, the tighter the matching}

\item{kernel.width}{the width of the deformation kernel for individual objects}

\item{plotting}{whether to plot the 3D objects in question ot help you decide on the above}

\item{...}{additional arguments eventually passed by methods}

\item{object.type}{type of object to be deformed. PointCloud,
OrientedPolyLine, NonOrientedPolyLine, OrientedSurfaceMesh and
NonOrientedSurfaceMesh}

\item{no.iterations}{how many iterations to go through before terminating the registration attempt}

\item{sparsity}{prior see deformetrica website for details}

\item{The}{path for the deformetrica app on your machine}
}
\value{
A list of deofrmed mesh3D objects, with the deformation stored in its attributes
}
\description{
Run a Deformatrica sparseMatching3 deformation on 3D objects
}
\references{
See \url{http://www.deformetrica.org/?page_id=189} for details of
  the \bold{sparseMatching3} command line tool.
}

