% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrisel1.R
\name{symmetrisel1}
\alias{symmetrisel1}
\alias{symmetrisel1.default}
\alias{symmetrisel1.neuron}
\alias{symmetrisel1.neuronlist}
\title{Symmetrise the L1 CNS space}
\usage{
symmetrisel1(x,
  symtrans = readRDS(system.file("extdata/symmetry_transform.rds", package =
  "deformetricar")), ...)

\method{symmetrisel1}{default}(x,
  symtrans = readRDS(system.file("extdata/symmetry_transform.rds", package =
  "deformetricar")), ...)

\method{symmetrisel1}{neuron}(x,
  symtrans = readRDS(system.file("extdata/symmetry_transform.rds", package =
  "deformetricar")), ...)

\method{symmetrisel1}{neuronlist}(x,
  symtrans = readRDS(system.file("extdata/symmetry_transform.rds", package =
  "deformetricar")), ...)
}
\arguments{
\item{x}{a set of 3D coordinates (Nx3 matrix), a neuron, neuronlist, dotprops, hxsurf or mesh3d object.
currently point replacement is only supported for neuron and neuronlist objects.}

\item{...}{additional arguments passed to: \code{\link{shootflow}} \code{\link{apply.mirror.affine}}}

\item{regdir}{Path to a tps transformation, generated by the Morpho package.Default symmetrises the larval l1 CNS, scaled in micrometers.}
}
\value{
a set of 3D cordinates (or a neuron/neuronlist object if that was given as input), that has undergone
symmetrisation
}
\description{
Symmetrise the L1 CNS space
}

