% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroraffine.R
\name{trafoicpmat}
\alias{trafoicpmat}
\title{Get affine transformation matrices}
\usage{
trafoicpmat(x, y, iterations, mindist = 1e+15, subsample = NULL,
  type = c("rigid", "similarity", "affine"), ...)
}
\arguments{
\item{x}{moving landmarks}

\item{y}{target landmarks}

\item{iterations}{integer number of iterations}

\item{mindist}{restrict valid points to be within this distance}

\item{subsample}{use a subsample determined by kmean clusters to speed up computation}

\item{type}{character: select the transform to be applied, can be "rigid","similarity" or "affine"}

\item{...}{additional arguments eventually passed to methods}
}
\value{
a data frame containing the transformed object, and the transformation matrices describing the performed
transformation
}
\description{
A modification of the icpmat function from the package Morpho. This function matches
two landmark configurations using iteratively closest point search.
}

