% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteVTK.R
\name{transform.vtk}
\alias{transform.vtk}
\title{Transform a .vtk object}
\usage{
\method{transform}{vtk}(vtk, transformations)
}
\arguments{
\item{vtk}{path to vtk object}

\item{transformations}{transformation(s)}
}
\description{
Transform a .vtk object
}

