% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroraffine.R
\name{transform3dpoints}
\alias{transform3dpoints}
\title{Transform 3D coordinates using a list of transformation matrices}
\usage{
transform3dpoints(positions, transformations, ...)
}
\arguments{
\item{positions}{3D coodinates}

\item{transformations}{list of transformation matrices}

\item{...}{additional arguments eventually passed to methods}
}
\value{
the transformed 3D coordinates
}
\description{
Transform 3D coordinates using a list of transformation matrices
}

