% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteVTK.R
\name{transform_vtk}
\alias{transform_vtk}
\title{Transform a .vtk object}
\usage{
transform_vtk(vtk, transformations)
}
\arguments{
\item{vtk}{path to vtk object}

\item{transformations}{transformation(s)}
}
\description{
Transform a .vtk object
}
