% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteVTK.R
\name{write.vtk.light}
\alias{write.vtk.light}
\title{Write a mesh3d object, removing internal points}
\usage{
write.vtk.light(mesh3d, alpha = 2, filename)
}
\arguments{
\item{mesh3d}{a mesh3d object}

\item{Whereto}{save the object}

\item{Value}{of alpha to be used with alphashape3d::ashape3d() to re-define the surface}
}
\description{
Write a mesh3d object, removing internal points
}

