% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformetricar.R
\docType{package}
\name{deformetricar-package}
\alias{deformetrica}
\alias{deformetricar,}
\alias{deformetricar-package}
\title{R wrapper for elements of the Deformetrica registration tool}
\description{
\bold{deformetricar} wraps the Deformetrica tool for statistical analysis of
2D and 3D shape data.
}
\section{Package Options}{
 The following options can be set to specify default
  behaviour.

  \itemize{

  \item{\code{deformeticar.bindir}}{ Location of Deformetrica binaries}. You
  may wish to set this in your \code{\link[base]{Rprofile}} if you do not
  have deformeticra installed in a standard location.

  }
}
\examples{
# Show state of deformetricar package options
options()[grep('^deformetricar', names(options()))]
}
\references{
See \url{http://www.deformetrica.org/}
}
\seealso{
\code{\link{shootflow}}
}
\keyword{package}

