% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherside.R
\name{otherside}
\alias{otherside}
\title{Apply a list of mirroring, affine and warping registrations to an object in 3D space}
\usage{
otherside(x, ...)
}
\arguments{
\item{x}{a set of 3D coordinates (Nx3 matrix), a neuron, neuronlist, dotprops, hxsurf, igraph or mesh3d object.
currently point replacement is only supported for neuron and neuronlist objects.}

\item{...}{additional arguments passed to: \code{\link{shootflow}} \code{\link{apply.mirror.affine}}}
}
\value{
a set of 3D cordinates (or a neuron/neuronlist object if that was given as input), that has undergone sequential mirroring
affine and warping (via Deformetrica, see http://www.deformetrica.org/?page_id=232) transformations. The default is to flip points in
the l1 larval central nervous system from one side to the other, and apply a previousy generated Deformetrica deformation based on the
registering of a set of flipped, large neuroanatomical regions unto their un-flipped counterparts.
}
\description{
Apply a list of mirroring, affine and warping registrations to an object in 3D space
}
\seealso{
\code{\link{shootflow}} \code{\link{apply.mirror.affine}}
}

