% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootflow.R
\name{shootflow}
\alias{shootflow}
\alias{shootflow.default}
\alias{shootflow.neuron}
\alias{shootflow.neuronlist}
\title{Apply deformetrica registration to a set of 3D points (ShootAndFlow3)}
\usage{
shootflow(x, ...)

\method{shootflow}{neuron}(x, ...)

\method{shootflow}{neuronlist}(x, ...)

\method{shootflow}{default}(x, regdir = system.file("extdata/reg_output/",
  package = "deformetricar"), kernel.width = 5,
  object.type = c("PointCloud", "OrientedPolyLine", "NonOrientedPolyLine",
  "OrientedSurfaceMesh", "NonOrientedSurfaceMesh"), ...)
}
\arguments{
\item{x}{object to be transformed (for \code{shootflow.default} method an Nx3
matrix of 3D coordinates)}

\item{...}{additional arguments eventually passed by methods
\code{shootflow.default}}

\item{regdir}{Path to directory containing deformetrica registration}

\item{kernel.width}{the width of the deformation kernel. The larger the more rigid
the deformation. The smaller, the more local variations of the space is allowed.}

\item{object.type}{type of object to be deformed. PointCloud, OrientedPolyLine, NonOrientedPolyLine,
OrientedSurfaceMesh and NonOrientedSurfaceMesh}
}
\value{
matrix of the same dimensions as \code{x}
}
\description{
Apply a Deformetrica deformation using calculated momentum vectors and control points (in files MOM_final.txt
and CP_final.txt respectively), by calling the executable ShootAndFlow3. This requires installation of Deformetrica version 2.1 or later (preferably in Applications).
}
\references{
See \url{http://www.deformetrica.org/?page_id=232} for details of
  the \bold{ShootAndFlow3} command line tool.
}

