% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteVTK.R
\name{write.vtk}
\alias{write.vtk}
\title{Write VTK file}
\usage{
write.vtk(filename, title = filename, points, polygons = NULL,
  normals = NULL, datatype = c("float", "double"))
}
\arguments{
\item{filename}{Path to output file}

\item{title}{Title of the .vtk file (defaults to filename)}

\item{points}{3D coordinates (Nx3 matrix)}

\item{polygons}{Triangle indices (Mx3 matrix)}

\item{normals}{Normals (Nx3 matrix)}

\item{datatype}{.vtk datatype (defaults to float)}
}
\description{
Write VTK file
}

